Removed rounding difference between "Best position" and "Position Rating", fixes #82.
git-svn-id: file:///home/coren/project/backup_ho/trunk/Core@519 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ src/java/de/hattrickorganizer/gui/templates/ColorLabelEntry.java
@@ -251,7 +251,7 @@ public class ColorLabelEntry extends TableEntry {
                            boolean farbeInvertieren, int nachkommastellen) {
     	m_iAusrichtung = SwingConstants.RIGHT;
         createComponent();
-    	setValueAsText((double)zahl, bg_color, currencyformat, farbeInvertieren, 
+    	setValueAsText(zahl, bg_color, currencyformat, farbeInvertieren, 
     			nachkommastellen, true);
     }
